+++ firefox-ios/Client/Frontend/Browser/BrowserViewController/Extensions/BrowserViewController+URLBarDelegate.swift
@@ -311,7 +311,8 @@ extension BrowserViewController: URLBarDelegate {
         }
 
         searchController?.searchQuery = text
-        searchController?.interactionType = .refined
+        searchController?.searchTelemetry?.searchQuery = text
+        searchController?.searchTelemetry?.interactionType = .refined
         searchLoader?.setQueryWithoutAutocomplete(text)
     }
 
@@ -326,18 +327,9 @@ extension BrowserViewController: URLBarDelegate {
             theme: self.themeManager.currentTheme
         )
         searchController?.searchQuery = text
+        searchController?.searchTelemetry?.searchQuery = text
         searchLoader?.query = text
-        searchController?.interactionType = determineInteractionType(for: text)
-    }
-
-    func determineInteractionType(for text: String) -> TelemetryWrapper.EventExtraKey.UrlbarTelemetry.Interaction {
-        guard let savedQuerry = searchController?.savedQuery else { return .typed }
-        if text.count - savedQuerry.count == 1 {
-            return .typed
-        } else if text.count - savedQuerry.count > 1 {
-            return .pasted
-        }
-        return .typed
+        searchController?.searchTelemetry?.determineInteractionType()
     }
 
     func urlBar(_ urlBar: URLBarView, didSubmitText text: String) {
