+++ firefox-ios/Client/Telemetry/SearchTelemetry.swift
@@ -5,6 +5,7 @@
 import Foundation
 import WebKit
 import Glean
+import Storage
 
 // Search Partner Codes
 // https://docs.google.com/spreadsheets/d/1HMm9UXjfJv-uHhGU1pJlbP4ILkdpSD9w_Fd-3yOd8oY/
@@ -29,11 +30,95 @@ enum SearchEngine: String, CaseIterable {
     case none
 }
 
+// all the events explained: 
+// https://docs.google.com/document/d/1FFWyxU_uNgjJQATeNbNQKRiSpMztDJKsGc09McUqE0A/edit?pli=1
+// QA about the search bar telemetry:
+// https://docs.google.com/document/d/1Cb4oEoGPTmDCPf0DCbEHhYOS1DRKGRjemsrprhuziKY/edit
+enum SearchTelemetryValues {
+    enum Sap: String {
+        case urlbar
+        case urlbarNewtab = "urlbar_newtab"
+    }
+
+    enum SearchMode: String {
+        case actions
+        case bookmarks
+        case history
+        case tabs // the only one that's valid for iOS at the moment
+        case unknown
+        case inactive = "" // Empty string for inactive search mode
+    }
+
+    enum Reason: String {
+        case pause
+    }
+
+    enum Interaction: String {
+        case typed
+        case pasted
+        case returned
+        case restarted
+        case refined
+        case persistedSearchTerms = "persisted_search_terms"
+        case persistedSearchTermsRestarted = "persisted_search_terms_restarted"
+        case persistedSearchTermsRefined = "persisted_search_terms_refined"
+    }
+
+    enum Groups: String {
+        case heuristic
+        case adaptiveHistory = "adaptive_history"
+        case searchHistory = "search_history"
+        case searchSuggest = "search_suggest"
+        case topPick = "top_pick"
+        case topSite = "top_site"
+        case remoteTab = "remote_tab"
+        case general
+        case suggest
+    }
+
+    enum Results: String {
+        case unknown
+        case bookmark
+        case history
+        case keyword
+        case searchEngine = "search_engine"
+        case searchSuggest = "search_suggest"
+        case searchHistory = "search_history"
+        case url
+        case action
+        case tab
+        case remoteTab = "remote_tab"
+        case tabToSearch = "tab_to_search"
+        case topSite = "top_site"
+        case suggestSponsor = "suggest_sponsor"
+        case suggestNonSponsor = "suggest_non_sponsor"
+    }
+}
+
 class SearchTelemetry {
     var code: String = ""
     var provider: SearchEngine = .none
     var shouldSetGoogleTopSiteSearch = false
     var shouldSetUrlTypeSearch = false
+    private var tabManager: TabManager
+
+    var interactionType: SearchTelemetryValues.Interaction = .typed
+    var impressionTelemetryTimer: Timer?
+
+    var remoteClientTabs = [ClientTabsSearchWrapper]()
+    var filteredOpenedTabs = [Tab]()
+    var filteredRemoteClientTabs = [ClientTabsSearchWrapper]()
+    var suggestions: [String]? = []
+    var firefoxSuggestions = [RustFirefoxSuggestion]()
+    var searchHighlights = [HighlightItem]()
+
+    var data: Cursor<Site> = Cursor<Site>(status: .success, msg: "No data set")
+    var searchQuery: String = ""
+    var savedQuery: String = ""
+
+    init(tabManager: TabManager) {
+        self.tabManager = tabManager
+    }
 
     // MARK: Searchbar SAP
 
@@ -104,6 +189,174 @@ class SearchTelemetry {
                 tab.urlType = .regular
             }
         }
+    }
+
+    // MARK: Impression Telemetry
+    func startImpressionTimer() {
+        impressionTelemetryTimer?.invalidate()
+        impressionTelemetryTimer = Timer.scheduledTimer(timeInterval: 1.0,
+                                                        target: self,
+                                                        selector: #selector(recordImpressionTelemetryEvent),
+                                                        userInfo: nil,
+                                                        repeats: false)
+    }
+
+    func stopImpressionTimer() {
+        impressionTelemetryTimer?.invalidate()
+    }
+
+    @objc
+    func recordImpressionTelemetryEvent() {
+        guard let tab = tabManager.selectedTab else { return }
+        let reasonKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.reason.rawValue
+        let reason = SearchTelemetryValues.Reason.pause.rawValue
+
+        let sapKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.sap.rawValue
+        let sap = checkSAP(for: tab).rawValue
+
+        let interactionKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.interaction.rawValue
+        let interaction = interactionType.rawValue
+
+        let searchModeKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.searchMode.rawValue
+        let searchMode = SearchTelemetryValues.SearchMode.tabs.rawValue
+
+        let nCharsKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.nChars.rawValue
+        let nChars = Int32(searchQuery.count)
+
+        let nWordsKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.nWords.rawValue
+        let nWords = numberOfWords(in: searchQuery)
+
+        let nResultsKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.nResults.rawValue
+        let nResults = Int32(numberOfSearchResults())
+
+        let groupsKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.groups.rawValue
+        let groups = listGroupTypes()
+
+        let resultsKey = TelemetryWrapper.EventExtraKey.UrlbarTelemetry.results.rawValue
+        let results = listResultTypes()
+
+        let extraDetails = [reasonKey: reason,
+                               sapKey: sap,
+                       interactionKey: interaction,
+                        searchModeKey: searchMode,
+                            nCharsKey: nChars,
+                            nWordsKey: nWords,
+                          nResultsKey: nResults,
+                            groupsKey: groups,
+                           resultsKey: results] as [String: Any]
+
+        TelemetryWrapper.recordEvent(category: .information,
+                                     method: .view,
+                                     object: .urlbarImpression,
+                                     extras: extraDetails)
+    }
+
+    func checkSAP(for tab: Tab?) -> SearchTelemetryValues.Sap {
+        guard let tab = tab else { return .urlbar }
+        if tab.isFxHomeTab || tab.isCustomHomeTab {
+            return .urlbarNewtab
+        }
+        return .urlbar
+    }
+
+    func determineInteractionType() {
+        if searchQuery.count - savedQuery.count == 1 {
+            interactionType = .typed
+        } else if searchQuery.count - savedQuery.count > 1 {
+            interactionType = .pasted
+        }
+    }
+
+    func numberOfWords(in string: String) -> Int32 {
+        let words = string.components(separatedBy: CharacterSet.whitespacesAndNewlines)
+        let filteredWords = words.filter { !$0.isEmpty }
+        return Int32(filteredWords.count)
+    }
+
+    func numberOfSearchResults() -> Int {
+        return suggestions?.count ?? 0 + data.count + searchHighlights.count
+        + filteredOpenedTabs.count + firefoxSuggestions.count
+        + filteredRemoteClientTabs.count
+    }
+
+    // Comma separated list of result types in order.
+    func listResultTypes() -> String {
+        var resultTypes: [String] = []
+
+        if let suggestionsCount = suggestions?.count, suggestionsCount > 0 {
+            resultTypes += Array(repeating: SearchTelemetryValues.Results.searchSuggest.rawValue,
+                                 count: suggestionsCount)
+        }
+
+        if !filteredOpenedTabs.isEmpty {
+            resultTypes += Array(repeating: SearchTelemetryValues.Results.tab.rawValue,
+                                 count: filteredOpenedTabs.count)
+        }
+
+        if !filteredRemoteClientTabs.isEmpty {
+            resultTypes += Array(repeating: SearchTelemetryValues.Results.remoteTab.rawValue,
+                                 count: filteredRemoteClientTabs.count)
+        }
+
+        for clientTab in data {
+            if let isBookmarked = clientTab?.bookmarked {
+                resultTypes.append(isBookmarked ?
+                                   SearchTelemetryValues.Results.bookmark.rawValue :
+                                    SearchTelemetryValues.Results.history.rawValue)
+            }
+        }
+
+        if !searchHighlights.isEmpty {
+            resultTypes += Array(repeating: SearchTelemetryValues.Results.searchHistory.rawValue,
+                                 count: searchHighlights.count)
+        }
+
+        for suggestion in firefoxSuggestions {
+            resultTypes.append(suggestion.isSponsored ?
+                               SearchTelemetryValues.Results.suggestSponsor.rawValue :
+                                SearchTelemetryValues.Results.suggestNonSponsor.rawValue)
+        }
+
+        return resultTypes.joined(separator: ",")
+    }
+
+    // Comma separated list of result groups in order, groups may be
+    // repeated, since the list will match 1:1 the results list, so we
+    // Can link each result to a group
+    func listGroupTypes() -> String {
+        var groupTypes: [String] = []
+
+        if !(suggestions?.isEmpty ?? true) {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.searchSuggest.rawValue,
+                                count: suggestions!.count)
+        }
+
+        if !filteredOpenedTabs.isEmpty {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.heuristic.rawValue,
+                                count: filteredOpenedTabs.count)
+        }
+
+        if !filteredRemoteClientTabs.isEmpty {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.remoteTab.rawValue,
+                                count: filteredRemoteClientTabs.count)
+        }
+
+        if !remoteClientTabs.isEmpty {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.general.rawValue,
+                                count: remoteClientTabs.count)
+        }
+
+        if !searchHighlights.isEmpty {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.searchHistory.rawValue,
+                                count: searchHighlights.count)
+        }
+
+        if !firefoxSuggestions.isEmpty {
+            groupTypes += Array(repeating: SearchTelemetryValues.Groups.suggest.rawValue,
+                                count: firefoxSuggestions.count)
+        }
+
+        return groupTypes.joined(separator: ",")
     }
 }
